[REFACTOR] obstacle speed changed from int to double

+++ src/main/java/it/unibo/project/game/model/api/Obstacle.java
@@ -32,5 +32,5 @@ public interface Obstacle extends Entity {
     /**
      * @return obstacle speed
      */
-    int speed();
+    double speed();
 }
